+++ packages/terra-core-docs/src/terra-dev-site/doc/button/example/IconButtonVariant.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import IconEdit from 'terra-icon/lib/icon/IconEdit';
-import { IconButton, IconTypes } from 'terra-button';
+import Button, { IconTypes } from 'terra-button';
 import classNames from 'classnames/bind';
 import styles from './ButtonDocCommon.module.scss';
 
@@ -8,8 +8,8 @@ const cx = classNames.bind(styles);
 
 const ButtonVariant = () => (
   <div>
-    <IconButton text="Action" variant="action" icon={<IconEdit a11yLabel="Edit" />} iconType={IconTypes.INFORMATIVE} className={cx('button')} />
-    <IconButton text="Utility" variant="utility" icon={<IconEdit />} iconType={IconTypes.DECORATIVE} className={cx('button')} />
+    <Button text="Action" variant="action" icon={<IconEdit a11yLabel="Edit" />} iconType={IconTypes.INFORMATIVE} className={cx('button')} />
+    <Button text="Utility" variant="utility" icon={<IconEdit />} iconType={IconTypes.DECORATIVE} className={cx('button')} />
   </div>
 );