+++ packages/terra-button/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Changed
   * Updated wdio screenshots due to functional-testing upgrade.
 * Added
+  * Added `IconTypes` for terra-button.
   * Added support for decorative icon buttons.
 
 ## 3.62.0 - (December 7, 2022)