+++ packages/terra-core-docs/CHANGELOG.md
@@ -6,12 +6,11 @@
   * Updated radio and checkbox tests to fix fusion theme tests.
 * Added 
   * Added Accessibility Guide for `terra-toggle`.
+  * Added A11y guide and examples for `terra-button`.
 
 * Updated
+  * Updated imports of decorative icons.
   * Updated jest & wdio tests.
-
-* Added
-  * Added A11y guide and examples for `terra-button`.
 
 ## 1.20.0 - (February 7, 2023)